Fix unnecessarily long multipart boundary length
Previously multipart boundary lengths were generated with 80
characters. Most web clients generate boundary lengths between 20 and
40 characters. Rack, and possibly other webserver interfaces, limit
boundaries to 70 characters. That caused exceptions to be raised in
the test environment when stubbing the Anthropic service, and created a
risk of boundary errors if the Anthropic API ever imposes a similar
constraint on boundary length in the future.

This solves the problem by reducing the boundary length from 80
characters to 40.