fix: always send filename in content-disposition headers (#126)
* Fix Content-Disposition part headers

File uploads were raising Anthropic::Errors::BadRequestError because
the Anthropic API was returning a 400 response with the message "Part
exceeded maximum size of 1024KB". That was because this gem was
sending Content-Disposition part headers without a "filename"
parameter. Although that parameter is not required by RFC, it is
required by many servers like Starlette (Anthropic) and Rack.

This changes Content-Disposition header generation so a "filename"
parameter is always given. This is what solves the file upload
problem.

This also ensures a "name" parameter is always given, because RFC 7578
requires it for the Content-Disposition header to be valid.

* Restore variable pinning in tests

* Change default file name to "file"

This changes the default file name from "upload" to match the Python
SDK.

* chore: test cases and clean up

---------

Co-authored-by: dogisgreat <git@bigly.dog>